+++ tests/testthat/test-cloning-inheritance.R
@@ -1,6 +1,4 @@
 test_that("Subclass can override superclass' cloneable property", {
-  msg <- "Sub and superclass have different cloneable properties."
-
   # superclass cloneable ---------------------
 
   Creature <- R6Class("Creature", cloneable = TRUE)
@@ -10,7 +8,7 @@ test_that("Subclass can override superclass' cloneable property", {
   expect_s3_class(sheep$clone(), "Sheep")
 
   Human <- R6Class("Human", inherit = Creature, cloneable = FALSE)
-  expect_message(human <- Human$new(), msg)
+  expect_message(human <- Human$new(), NA)
   expect_error(human$clone(), "attempt to apply non-function")
 
   # superclass non-cloneable  ---------------------
@@ -18,8 +16,8 @@ test_that("Subclass can override superclass' cloneable property", {
   Creature <- R6Class("Creature", cloneable = FALSE)
 
   Sheep <- R6Class("Sheep", inherit = Creature, cloneable = TRUE)
-  expect_message(sheep <- Sheep$new(), msg)
-  expect_s3_class(sheep$clone(), "Sheep")
+  expect_message(sheep <- Sheep$new(), "Subclass wants to allow cloning, but superclass has turned it off.")
+  expect_error(sheep$clone(), "attempt to apply non-function")
 
   Human <- R6Class("Human", inherit = Creature, cloneable = FALSE)
   expect_message(human <- Human$new(), NA)